Add files via upload
Our project is a User Registration System built using HTML, CSS, and JavaScript. It allows users to create an account by entering their details such as Name, Email, Password, and Confirm Password. The main focus of the project is on form validation to ensure that the user enters correct and valid information before submission. Thankyou